# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Belconnen United FC Performance Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      parameters:
        - name: teamId
          in: query
          schema: { type: integer }
        - name: seasonId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Create a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get a player
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
    patch:
      operationId: updatePlayer
      tags: [players]
      summary: Update a player
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"

  # ─── TEAMS ──────────────────────────────────────────────────────────────────
...
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get a team
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"

  # ─── SEASONS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Player:
      type: object
      properties:
        id:        { type: integer }
        name:      { type: string }
        position:  { type: ["string", "null"] }
        club:      { type: ["string", "null"] }
        dob:       { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, name, createdAt]
    PlayerInput:
      type: object
      properties:
        name:     { type: string }
        position: { type: ["string", "null"] }
        club:     { type: ["string", "null"] }
        dob:      { type: ["string", "null"] }
      required: [name]
    PlayerUpdate:
      type: object
      properties:
        name:     { type: string }
        position: { type: ["string", "null"] }
        club:     { type: ["string", "null"] }
        dob:      { type: ["string", "null"] }

    # ── Teams ─────────────────────────────────────────────────────────────────
    Team:
      type: object
      properties:
        id:               { type: integer }
        name:             { type: string }
        gender:           { type: string }
        ageGroup:         { type: string }
        analyticsEnabled: { type: boolean }
        clubName:         { type: string }
      required: [id, name, gender, ageGroup, analyticsEnabled, clubName]

    # ── Seasons ───────────────────────────────────────────────────────────────
...